% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_afib.R
\docType{data}
\name{dat_afib}
\alias{dat_afib}
\title{Stroke prevention therapies for people with atrial fibrillation}
\format{
A data frame with 49 rows and 14 variables:
\describe{
\item{\code{TE}}{The observed relative treatment effects.}
\item{\code{seTE}}{The standard error of the observed relative treatment effects.}
\item{\code{studlab}}{The study labels.}
\item{\code{treat1}}{The treatment code of the comparator in the observed relative
effect.}
\item{\code{treat2}}{The treatment code of the intervention in the observed
relative effect.}
\item{\code{event1}}{The number of participants assigned \code{treat1} and experienced
a stroke during follow-up.}
\item{\code{time1}}{The total number of months at risk for participants assigned
\code{treat1}.}
\item{\code{event2}}{The number of participants assigned \code{treat2} and experienced
a stroke during follow-up.}
\item{\code{time2}}{The total number of months at risk for participants assigned
\code{treat2}.}
\item{\code{stroke}}{The proportion of participants in the study who previously
experienced a stroke before randomization.}
\item{\code{class1_detail}}{The drug class of the treatment in \code{treat1}.}
\item{\code{class2_detail}}{The drug class of the treatment in \code{treat2}.}
\item{\code{class1_simple}}{\code{treat1} categorized as "Control" or "Active".}
\item{\code{class2_simple}}{\code{treat2} categorized as "Control" or "Active".}
}
}
\source{
Cooper et al. (2009). Addressing between-study heterogeneity and inconsistency in mixed treatment comparisons: Application to stroke prevention treatments in individuals with non-rheumatic atrial fibrillation, Statistics in Medicine 28:1861-1881. \link{doi:10.1002/sim.3594}. Adopted from gemtc::atrialFibrillation.

data(dat_afib)
}
\usage{
dat_afib
}
\description{
Contrast-level data.
}
\keyword{datasets}
