% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_parkinsons.R
\docType{data}
\name{dat_parkinsons}
\alias{dat_parkinsons}
\title{Parkinson's NMA dataset}
\format{
A data frame with 9 rows and 11 columns:
\describe{
\item{\code{studlab}}{The study labels.}
\item{\code{treat1}}{The treatment code of the comparator in the observed relative
effect.}
\item{\code{treat2}}{The treatment code of the intervention in the observed
relative effect.}
\item{\code{mean1}}{.}
\item{\code{sd1}}{.}
\item{\code{n1}}{.}
\item{\code{mean2}}{.}
\item{\code{sd2}}{.}
\item{\code{n2}}{.}
\item{\code{TE}}{The observed relative treatment effects.}
\item{\code{seTE}}{The standard error of the observed relative treatment effects.}
}
}
\source{
Dias, S., Welton, N. J., Sutton, A. J., and Ades, A. E. (2011). A generalised linear modelling framework for pairwise and network meta-analysis of randomised controlled trials. In NICE DSU Technical Support Document 2. National Institute for Health and Care Excellence (NICE).

data(dat_parkinsons)
}
\usage{
dat_parkinsons
}
\description{
Contrast-level data
}
\keyword{datasets}
