% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_gfl.R
\name{find_best_gfl}
\alias{find_best_gfl}
\title{Determine the GFL problem which has the best solution}
\usage{
find_best_gfl(mod_sum_list, mod_sum_df, penfit = "AICc")
}
\arguments{
\item{mod_sum_list}{a list of all model summaries obtained from
\code{tabulate_solution_path()}.}

\item{mod_sum_df}{a data frame of all model summaries obtained from
\code{tabulate_solution_path()}.}

\item{penfit}{a string indicating which penalized fit measure will be used to
select the models: one of "AICc" or "BICc".}
}
\value{
A numeric value(s) indicating the index (indicies) of the GFL
problem(s) whose solution contains the model with the best penalized fit.
}
\description{
\code{find_best_gfl()} returns the index/indicies of the GFL problem for which the
solution contains the the model with the best penalized fit overall.
}
