% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groupings.R
\name{get_groupings}
\alias{get_groupings}
\title{Determine the pooled parameter groupings for a given lambda}
\usage{
get_groupings(par_est, par_groups = NULL, pool_format = TRUE)
}
\arguments{
\item{par_est}{a vector of (T-1) + P*(T-1) parameter estimates (relative
effects vs. reference treatment + treatment-covariate interactions) in a
network of T treatments, where P covariates have been adjusted for.}

\item{par_groups}{a list of lists of grouped indicies corresponding to sets
of parameters to be pooled within but not between.}

\item{pool_format}{a logical indicating whether pooled parameters should be
presented as a character string (TRUE) or list (FALSE). Default is TRUE.}
}
\value{
A list of the parameter poolings and the labels of parameters
belonging to each group.
}
\description{
\code{get_groupings()} groups pooled parameters if their estimates are (nearly)
equal, as determined with \code{all.equal()}.
}
