% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_penalty_matrix.R
\name{get_penalty_matrix}
\alias{get_penalty_matrix}
\title{Determine the penalty matrix to implement GFL within parameter groups}
\usage{
get_penalty_matrix(X, x_cov = NULL, gamma = 1)
}
\arguments{
\item{X}{a design matrix corresponding to the observed relative effects
in \code{y} inputted into \code{solve_gflnma()}, which may be obtained from
\code{get_design_matrix()}. This matrix may also be a binded matrix of the
design matrices for the relative effects and the interaction terms, the
latter of which may be obtained from \code{get_covariate_design_matrix()}.}

\item{x_cov}{an optional matrix containing the observed study-level covariate
values for P covariates, where the number of rows is equal to the length of
the inputted \code{y} in \code{solve_gflnma()} and the number of columns is equal to
P, the number of covariates. This should be specified if X includes
interaction terms}

\item{gamma}{a non-negative number indicating the ratio of penalization
factors (treatment-specific covariate interactions (b)/relative treatment
effects (d)). Default is 1.}
}
\value{
A matrix consisting of with (T-1) + P*(T-1) columns, where T is the
number of treatments in the network, and P is the number of covariates.
}
\description{
\code{get_penalty_matrix()} determines a matrix of the penalty terms for the
relative treatment effects and the treatment-covariate interaction terms (if
applicable), where the latter is scaled by the desired ratio of penalization
factors (\code{gamma}, see details for more information).
}
\details{
Currently designed for GFL-NMA where the relative treatment effects against a
reference treatmetn (d) and their pairwise differences are penalized, as well
as GFL-NMR which additionally (but separately) penalizes the treatment-
covariate interactions (b), as well as their pairwise differences.

Regarding the ratio of penalty factors: if gamma > 1, then the treatment-
covariate interactions will be penalized gamma times more than the treatment
effects. If gamma = 0, the treatment-covariate interactions are not
penalized. Treatment effects and their pairwise differences receive the same
penalty, while treatment-covariate interactions and their pairwise
differences receive gamma*that penalty.
}
