% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_solution_path.R
\name{plot_solution_path}
\alias{plot_solution_path}
\title{Plot the primal solution path}
\usage{
plot_solution_path(
  gfl_soln,
  mod_index = NULL,
  par_groups = NULL,
  gfl_sum = NULL,
  penfit = NULL,
  delta_penfit_lim = NULL,
  lambda_scale = 1,
  xlimits = NULL,
  ylimits = NULL,
  cex_lab = 2,
  cex_axis = 2,
  cex_text = 0.8
)
}
\arguments{
\item{gfl_soln}{an object of class \code{genlasso}.}

\item{mod_index}{an optional numeric value indicating the index of the
GFL-NMA problem to plot, if multiple problems were solved by
\code{solve_gflnma()} (e.g.,if \code{length(gamma)} > 1).}

\item{par_groups}{an optional list of lists of grouped indices corresponding
to sets of parameters (excluding reference) to be plotted separately
(e.g., d, b1, b2, ...).}

\item{gfl_sum}{an optional object of class \code{genlasso_summary} produced by the
\code{tabulate_solution_path()} function, which is used to determine the
critical lambda values for which parameters pool or unpool.}

\item{penfit}{an optional string indicating which penalized fit measure will
be used to select the models: one of "AIC", "AICc", "BIC", "BICc".}

\item{delta_penfit_lim}{an optional number indicating the range of
delta_\code{penfit} values corresponding to models the user is willing to
consider. Default is NULL.}

\item{lambda_scale}{a number indicating the desired rescaling power for the
presentation of lambda (i.e., lambda_rescaled = lambda ^ \code{lambda_scale}).
Default is 1.}

\item{xlimits}{numeric vector of length 2, giving the x coordinates range.}

\item{ylimits}{numeric vector of length 2, giving the y coordinates range.}

\item{cex_lab}{a numeric value indicating the expansion factor for the axis
titles.}

\item{cex_axis}{a numeric value indicating the expansion factor for the axis
labels.}

\item{cex_text}{a numeric value indicating the expansion factor for the
superimposed parameter labels.}
}
\value{
One or more plots of the estimated parameters against lambda (the
penalty factor for d).
}
\description{
\code{plot_solution_path()} returns a plot of the primal solution path for objects
of class \code{genlasso}. User must provide the \code{genlasso} object obtained from
\code{solve_gflnma()}.
}
\details{
To draw vertical dashed lines at critical lambda values for which the
parameters pool or unpool, the user must provide a \code{genlasso_summary} object
obtained from \code{tabulate_solution_path()} via the \code{gfl_sum} argument. The user
may draw additional lines at specific lambda values on an ad-hoc basis via
\code{abline()}.

The user also has the option to specify a critical region of \code{penfit} values
through \code{delta_penfit_crit}. See Kong et al. (2024) for recommended rules
of thumb. Additionally, the user has the option to specify which sets of
parameters should be plotted separately by specifying a listed of their
grouped indices through \code{par_groups}.
}
