% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gfl_data.R
\name{prep_gfl_data}
\alias{prep_gfl_data}
\title{Prepare contrast-level data for GFL-NMA or GFL-NMR}
\usage{
prep_gfl_data(
  dat,
  TE,
  studlab,
  seTE,
  treatment1,
  treatment2,
  se1 = NULL,
  sd1 = NULL,
  n1 = NULL,
  event1 = NULL,
  time1 = NULL,
  modtype = "ME"
)
}
\arguments{
\item{dat}{a data frame containing the observed data in each study.}

\item{TE}{the name of the column in \code{dat} that contains the estimates of the
relative treatment effects.}

\item{studlab}{the name of the column in \code{dat} that contains the study
labels.}

\item{seTE}{the name of the column in \code{dat} that contains the standard errors
of the relative treatment effects.}

\item{treatment1}{the name of the column in \code{dat} that contains the labels of
the comparator for each relative treatment effect in \code{TE} (\code{treatment2}
vs. \code{treatment1}).}

\item{treatment2}{the name of the column in \code{dat} that contains the labels of
the intervention for each relative treatment effect in \code{TE} (\code{treatment2}
vs. \code{treatment1}).}

\item{se1}{the name of the column in \code{dat} that contains the standard error
of the observed estimates in the comparator arm.}

\item{sd1}{the name of the column in \code{dat} that contains the standard
deviation of the observed outcomes in the comparator arm.}

\item{n1}{the name of the column in \code{dat} that contains the number of
participants in the comparator arm.}

\item{event1}{the name of the column in \code{dat} that contains the number of
events observed among participants in the comparator arm.}

\item{time1}{the name of the column in \code{dat} that contains the total person
time at risk among participants in the comparator arm.}

\item{mod_type}{a character string indicating if a fixed effect (FE), random
effects (RE), or multiplicative effects (ME) model is to be fitted.}
}
\value{
A list of
\itemize{
\item a data frame containing the required study-level data for GFL-NMA or
GFL-NMR,
\item a C x C covariance matrix for the observed relative effects, where C is
the number of rows in the outputted data frame,
\item a character string indicating the model assumed for the covariance
in the outputted data frame.
}
}
\description{
\code{prep_gfl_data()} determines the covariance of relative effects (vs. a common
comparator) in multi-arm studies and returns a list of a data frame that only
includes the data required by \code{solve_gflnma()}, as well as the variance-
covariance matrix required by \code{solve_gflnma()}, and a string indicating the
assumed type of effect for between-study heterogeneity.
}
\details{
User must specify if they want to fit a fixed effect (FE), random effects
(RE), or multiplicative effects (ME) model. This is critical for
determining the correct variance-covariance matrix.

\code{prep_gfl_data()} currently calculates covariance between contrasts within
studies using the variance of the outcome in the common comparator group via
one of the following:
\itemize{
\item \code{se1}, the standard error of the observed estimates in the comparator
group. If all pairwise relative effects are available, this argument may be
used to input the covariance via equation (4.12) of Dias et al. (2018).
\item \code{sd1} and \code{n1}, the standard deviation and sample size of the observed
outcomes in the comparator group.
\item \code{event1} and \code{time1}, the number of observed events and the person time
at risk in the comparator groups.
\item \code{event1} and \code{n1}, the number of observed events and the number of
participants in the comparator groups.
}

\code{prep_gfl_data()} currently calculates the random effect of between-study
heterogeneity based on a random effects model fitted via
\code{netmeta::netmeta()}. We recommend using multiplicative effects GFL-NMR over
random effects GFL-NMR.
}
\references{
Dias S, Ades AE, Welton NJ, Jansen JP, Sutton AJ. Network meta-
analysis for decision-making. John Wiley & Sons; 2018.
}
