% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gfl_params.R
\name{recode_gfl_params}
\alias{recode_gfl_params}
\title{Recode parameters based on configuration suggested by selected GFL model}
\usage{
recode_gfl_params(param_est, gfl_dat, par_groups, trt_codes)
}
\arguments{
\item{param_est}{a vector containing the parameter estimates of the desired
pooled model.}

\item{gfl_dat}{data frame from \code{prep_gfl_data}.}

\item{par_groups}{a list of parameter indices (excluding reference) in the
full model.}

\item{trt_codes}{a data frame containing the original treatment
codes (\code{code}) and names (\code{trt}). This is required if the original data set
contains the treatment names only.}
}
\value{
a data frame that contains data in \code{gfl_dat}, along with appended
grouped parameter codes.
}
\description{
\code{recode_gfl_params()} returns a matrix containing codes for the pooled
parameter groups as determined by a selected GFL-NMR model.
}
