% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_treatments.R
\name{recode_treatments}
\alias{recode_treatments}
\title{Recode the treatments in sequential order}
\usage{
recode_treatments(treatment1, treatment2, ref)
}
\arguments{
\item{treatment1}{a numeric (or character) vector with the same length as
\code{treatment2}, which contains the codes (or names) of the comparator for
each observed relative treatment effect (\code{treatment2} vs. \code{treatment1}).}

\item{treatment2}{a numeric (or character) vector with the same length as
\code{treatment1}, which contains the codes (or names) of the intervention for
each observed relative treatment effect (\code{treatment2} vs. \code{treatment1}).}

\item{ref}{a single character string indicating the code (or name) of the
network reference treatment.}
}
\value{
A matrix of \code{length(treatment1)} rows and two columns.
}
\description{
\code{recode_treatments()} returns a matrix containing new codes for the
inputted \code{treatment1} and \code{treatment2} which correspond to some observed
(relative) treatment effects. The new codes ensure the reference treatment
(\code{ref}) is coded as treatment 1, and the remaining treatments are coded from
2 to T in numerical (or alphabetical) order of their original codes (or
names), where T is the number of treatments in the network.
}
