% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_gflnma.R
\name{solve_gflnma}
\alias{solve_gflnma}
\title{Solve one or more GFL-NMA or GFL-NMR models}
\usage{
solve_gflnma(
  y,
  var_cov,
  x_cov = NULL,
  treatment1,
  treatment2,
  ref,
  class1 = NULL,
  class2 = NULL,
  ref_class = NULL,
  fit_full = FALSE,
  minlam = 0,
  gamma = 0,
  eps = 1e-04,
  center = TRUE,
  rescale = TRUE
)
}
\arguments{
\item{y}{a numeric vector containing the observed relative treatment effects
vs. a study-specific common treatment for each study.}

\item{var_cov}{a block-diagonal matrix of each study's variance-covariance
matrix corresponding to observed relative treatment effects in \code{y}, where
the number of rows and number of columns both equal the length of the \code{y}.}

\item{x_cov}{a matrix containing the observed study-level covariate values
for P covariates, where the number of rows is equal to the length of \code{y}
and the number of columns is equal to P, the number of covariates.}

\item{treatment1}{a numeric (or character) vector with the same length as
\code{y}, which contains the codes (or names) of the comparator for each
observed relative treatment effect (\code{treatment2} vs. \code{treatment1}).}

\item{treatment2}{a numeric (or character) vector with the same length as
\code{y}, which contains the codes (or names) of the intervention for each
observed relative treatment effect (\code{treatment2} vs. \code{treatment1}).}

\item{ref}{a single character string indicating the code (or name) of the
network reference treatment.}

\item{class1}{an optional numeric (or character) vector with the same length
as \code{y}, which contains the codes (or names) of \code{treatment1}'s class for
each observed relative treatment effect (\code{class2} vs. \code{class1}). Only input
if the treatment-covariate terms should be shared within classes.}

\item{class2}{an optional numeric (or character) vector with the same length
as \code{y}, which contains the codes (or names) of \code{treatment2}'s class for
each observed relative treatment effect (\code{class2} vs. \code{class1}). Only input
if the treatment-covariate terms should be shared within classes.}

\item{ref_class}{a single character string indicating the code (or name) of
the network reference treatment class.  Only input if the treatment-
covariate terms should be shared within classes.}

\item{fit_full}{a logical value indicating if the full model should be fitted
(TRUE) or not (FALSE, default).}

\item{minlam}{a numeric variable indicating the value of tuning parameter for
the d's at which the solution path should terminate. Default is 0.}

\item{gamma}{a vector of numeric values > 0 indicating the desired ratio of
the penalty for interaction terms (beta's) to the penalty for the relative
treatment effects (d's). Default is 0 to only penalize the relative
treatment effects. If only interaction terms are to be penalized, input
\code{Inf} (no quotes).}

\item{eps}{a vector of numeric values > 0 indicating the desired
multiplier(s) for the ridge penalty, which will be implemented if there are
some parameters for which there is no direct nor indirect evidence in the
data. Default is 0.0001.}

\item{center}{either a logical value or a numeric vector  of equal length to
the number of covariates indicating whether covariates should be centered
by their means (TRUE, the default), centered by specific values supplied in
the numeric vector, or not at all (FALSE).}

\item{rescale}{either a logical value or a numeric vector  of equal length to
the number of covariates indicating whether covariates should be scaled
by their standard deviation (TRUE, the default), scaled by specific values}
}
\value{
A nested list of
\itemize{
\item \code{solution}: a nested list of multiple \code{genlasso::genlasso} objects.
\item \code{par_groups}: a list of lists of parameter indices (excluding reference).
}
}
\description{
\code{solve_gflnma()} constructs one or more penalty matrix based on inputted
\code{gamma} and calls on \code{genlasso::genlasso()} to compute the solution path of
the generalized fused lasso problem(s).
}
