% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{tabulate_single_solution_path}
\alias{tabulate_single_solution_path}
\title{Produce a table summarizing critical points along a single solution path}
\usage{
tabulate_single_solution_path(
  x,
  lambda_scale,
  lambda_scale_digits,
  lambda_critical,
  full_rss,
  penfit,
  penfit_threshold,
  penfit_digits,
  par_groups
)
}
\arguments{
\item{x}{an object of class "\code{genlasso}".}

\item{lambda_scale}{a number indicating the desired rescaling power for the
presentation of lambda (i.e., lambdaRescaled = lambda ^ lambda_scale).
Default is 1.}

\item{lambda_scale_digits}{a number indicating desired number of digits for
rescaled lambda.}

\item{lambda_critical}{a character string, one of "fused", "all", or "best",
indicating whether vertical dashed lines should be drawn at lambdas for
which treatments are fused (or unfused), all hitting and leaving points on
the dual path boundary (see Tibshirani and Taylor 2016), or the best
fitting model based on AICc (i.e., the one with the smallest AICc). Default
is "fuse".}

\item{full_rss}{a number indicating the RSS of the full model (if estimable).}

\item{penfit}{a string indicating which penalized fit measure will be used to
select the models: one of "AIC", "AICc" (default), "BIC", "BICc".}

\item{penfit_threshold}{a number specifying the degrees of freedom threshold
for which the penalized fit measure should not be computed. See details.}

\item{penfit_digits}{a number indicating desired number of digits for the
penalized fit measures.}

\item{par_groups}{a list of lists of grouped indices corresponding to sets of
of parameters to be plotted separately.}
}
\value{
A data frame summarizing the df, lambda, RSS, AIC, AICc and treatment
groupings.
}
\description{
\code{tabulate_single_solution_path} calls on \code{genlasso::summary.genlasso} to
tabulate the degrees of freedom (df), the tuning parameter for the relative
treatment effects (lambda), and residual sum of squares (RSS), the
corresponding penalized model fit statistics (AIC, AICc, BIC, BICc), and the
parameter groupings at each critical lambda where parameters pool or
unpool.
}
\details{
In star networks with one study per comparison, the degrees of freedom (df)
equals the number of studies (M) when no treatments are pooled. The AIC and
BIC measures corrected for small sample size are minimized when df = M, and
so the full model will always be selected in these networks. Additionally,
the correction term for small sample size in AIC and BIC will be undefined
when df = M - 1. Thus, for these networks, we recommend not considering these
models. By default, the table will exclude these models in the summary
output. Alternatively, the user can specify \code{penfit_threshold} such that the
table will exclude models in the summary output for which
(df+1)/M >= \code{penfit_threshold}. If the user wants all models to be presented,
set \code{penfit_threshold} to be any number > (df+1)/M.
}
