% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_solution_path.R
\name{tabulate_solution_path}
\alias{tabulate_solution_path}
\title{Produce a table summarizing critical points along one or more solution paths}
\usage{
tabulate_solution_path(
  x,
  lambda_scale = 1,
  lambda_critical = "fused",
  full_rss = NA,
  penfit = "AICc",
  penfit_threshold = NULL,
  digits = c(3, 2, 2),
  par_groups = NULL
)
}
\arguments{
\item{x}{a list of one or more objects of class "genlasso".}

\item{lambda_scale}{a number indicating the desired rescaling power for the
presentation of lambda (i.e., lambdaRescaled = lambda ^ lambda_scale).
Default is 1.}

\item{lambda_critical}{a character string, one of "fused", "all", or "best",
indicating whether vertical dashed lines should be drawn at lambdas for
which treatments are fused (or unfused), all hitting and leaving points on
the dual path boundary (see Tibshirani and Taylor 2016), or the best
fitting model based on AICc (i.e., the one with the smallest AICc). Default
is "fuse".}

\item{full_rss}{a number indicating the RSS of the full model (in the case of
a connected network).}

\item{penfit}{a string indicating which penalized fit measure will be used to
select the models: one of "AIC", "AICc" (default), "BIC", "BICc".}

\item{penfit_threshold}{a number specifying the degrees of freedom threshold
for which the penalized fit measure should not be computed. See details.}

\item{digits}{a vector of numbers indicating desired number of digits for
lambda, rss, and penalized fit measures, respectively.}

\item{par_groups}{a list of lists of grouped indices corresponding to sets of
of parameters (e.g., d, b1, b2, ...).}
}
\value{
A list or data frame summarizing the degrees of freedom, lambda, RSS,
AIC, AICc, BIC, BICc, and parameter groupings.
}
\description{
\code{tabulate_solution_path()}...
}
