% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_solution_path.R
\name{plot_solution_path}
\alias{plot_solution_path}
\title{Plot the primal solution path along with critical points of interest}
\usage{
plot_solution_path(
  gfl_soln,
  gfl_sum,
  lambda_scale = 1,
  penfit = "AICc",
  delta_penfit_lim = NULL,
  par_groups = NULL,
  xlimits = NULL,
  ylimits = NULL
)
}
\arguments{
\item{gfl_soln}{an object of class \code{genlasso}.}

\item{gfl_sum}{an object of class \code{genlasso_summary} produced by the
\code{genlasso_summary_table()} function.}

\item{lambda_scale}{a number indicating the desired rescaling power for the
presentation of lambda (i.e., lambda_rescaled = lambda ^ \code{lambda_scale}).
Default is 1.}

\item{penfit}{a string indicating which penalized fit measure will be used to
select the models: one of "AIC", "AICc" (default), "BIC", "BICc".}

\item{delta_penfit_lim}{an optional number indicating the range of
delta_\code{penfit} values corresponding to models the user is willing to
consider. Default is NULL.}

\item{par_groups}{an optional list of lists of grouped indicies corresponding
to sets of parameters to be plotted separately (e.g., d, b1, b2, ...).}

\item{xlimits}{numeric vector of length 2, giving the x coordinates range.}

\item{ylimits}{numeric vector of length 2, giving the y coordinates range.}
}
\value{
One or more plots of the estimated parameters against lambda (the
penalty factor for d).
}
\description{
\code{plot_solution_path()} returns a plot of the primal solution path for objects
of class \code{genlasso}. User must provide the \code{genlasso} object obtained from
\code{solve_gflnma()} as well as a \code{genlasso_summary} object obtained from
\code{genlasso_summary_table()}.
}
\details{
The solution path will draw vertical dashed lines at critical lambda values
summarized in the \code{genlasso_summary} object. User may draw additional lines
at specific lambda values on an ad-hoc basis via \code{abline()}.

The user also has the option to specify a critical region of \code{penfit} values
through \code{delta_penfit_crit}. See Kong et al. (2024) for recommended rules
of thumb. Additionally, the user has the option to specify which sets of
parameters should be plotted separately by specifying a listed of their
grouped indices through \code{par_groups}.
}
